Make writeData and readData use size_t parameters

+++ src/util/include/util/serializing/ObjectEncoding.h
@@ -21,7 +21,7 @@ public:
 
 public:
     void addStr(const char* str) const;
-    virtual void addData(const void* data, int len) = 0;
+    virtual void addData(const void* data, size_t len) = 0;
 
     GString* getData();
